requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
streamlit>=1.35.0
python-dotenv>=1.0.0